pyproject.toml
[project]
name = "CyberAlchemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.6.4",
    "autogen-ext[azure,openai]>=0.6.4",
    "azure-identity>=1.23.1",
    "streamlit>=1.46.1",
    "python-dotenv>=1.1.1",
]
